data: Move live text lifecycle detection
Text lifecycle classification is pure core logic, while empty-path
detection reads repository state.

The detector lived beside pure lifecycle decisions, so core reached into
editor loaders and Git paths.

This commit moves the repository-bound detector to data, updates live
callers, and leaves core with buffer-based classification only.

Core modules now compile without editor imports, which gives lower-layer
parsing and lifecycle code a narrower dependency surface.
