+++ src/git_stage_batch/commands/include.py
@@ -47,7 +47,8 @@ from ..core.line_selection import (
     write_line_ids_file,
 )
 from ..core.models import BinaryFileChange, GitlinkChange, RenameChange, TextFileDeletionChange
-from ..core.text_lifecycle import TextFileChangeType, detect_empty_text_lifecycle_change
+from ..core.text_lifecycle import TextFileChangeType
+from ..data.text_lifecycle_detection import detect_empty_text_lifecycle_change
 from ..data.hunk_tracking import (
     apply_line_level_batch_filter_to_cached_hunk,
     fetch_next_change,
